refactored to remove the BL header subfolder.
